moved getExcelFiles function to util file
