+++ server/util/index.php
@@ -14,4 +14,16 @@
     }
     return true;
   }
+
+  function getXLSXFiles($path) {
+    if (is_dir($path)) {
+        $res = array();
+        foreach (array_filter(glob($path ."*.xlsx"), 'is_file') as $file) {
+            array_push($res, str_replace($path, "", $file));                
+        }
+        return $res;
+    }
+    return false;
+  }
+
 ?>
